backend/requirements.txt
flask
flask-cors
pytorch